on the way to get it running on other envs

+++ Analyze_Grid/setup_area.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 
 PS3="Your choice:  "
-source /cvmfs/cms.cern.ch/cmsset_default.sh
-export slc6_amd64_gcc491
-eval `scramv1 runtime -sh`
+#source /cvmfs/cms.cern.ch/cmsset_default.sh
+#export slc6_amd64_gcc491
+#eval `scramv1 runtime -sh`
 
 ###### variables needed
 username=$(whoami)
@@ -34,7 +34,7 @@ elif [ $n_analyzer_list -eq 1 ]; then
 else
     echo
     echo "Please select a number corresponding to the Analyzer you wish to use:"
-    select tmp_area in $analyzer_list; do 
+    select tmp_area in $analyzer_list; do
         if [ -z "$tmp_area" ]; then
 	    echo "Not a valid number"
 	else
@@ -45,27 +45,18 @@ else
 fi
 
 ##### download pyslurm stuff
-git clone https://github.com/PySlurm/pyslurm
-cd pyslurm
-tmp_slurm=$(which sbatch)
-slurm=$(echo $tmp_slurm | sed -rn 's|(.*slurm).*|\1|gp')
-python setup.py build --slurm=$slurm
-python setup.py install --user
+if [ -z "$(python -m pip list 2>/dev/null | grep pyslurm)" ]; then
+    git clone https://github.com/PySlurm/pyslurm
+    cd pyslurm
+    git checkout 16.05.5
+    tmp_slurm=$(which sbatch)
+    slurm=$(echo $tmp_slurm | sed -rn 's|(.*slurm).*|\1|gp')
+    python setup.py build --slurm=$slurm
+    python setup.py install --user
 
-IFS=':'
-working_path=''
-for path in $PYTHONPATH; do
-    if [ -d $path/pyslurm ]; then
-	working_path=$path
-	break
-    fi
-done
+    cd $analyze_grid
 
-cd $analyze_grid
-cp defaults/pyslurm_tmp.py $working_path/pyslurm
+    rm -rf pyslurm
+fi
 
-rm -rf pyslurm
-rm pyslurm_tmp.py
-
-###### move default files to main directory
-sed -e 's@ANALYZER_AREA@'"$analyzer_area"'@g' < default/run_slurm.slurm >run_slurm.slurm
+sed -e 's@ANALYZER_AREA@'"$analyzer_area"'@g' < defaults/run_slurm.slurm >run_slurm.slurm
